+++ arithmetic/recursive_decent_parser_test.go
@@ -1,4 +1,4 @@
-package parser
+package arithmetic
 
 import (
 	"fmt"
